Add outlets for the username and password text fields.
Set the window controller as delegate. Attempt to find credentials in the keychain when loading the window. Attempt to find credentials in the keychain anytime the username text field ends editing and the password text field is empty.

Save the last username that was used so that we can query the right account from the keychain on next launch.